+++ packages/react/src/index.ts
@@ -11,6 +11,10 @@ import {
 	computed,
 	batch,
 	effect,
+	action,
+	type Model,
+	type ModelConstructor,
+	createModel,
 	Signal,
 	type ReadonlySignal,
 	untracked,
@@ -20,6 +24,7 @@ import {
 	useSignal,
 	useComputed,
 	useSignalEffect,
+	useModel,
 } from "@preact/signals-react/runtime";
 
 export {
@@ -27,11 +32,16 @@ export {
 	computed,
 	batch,
 	effect,
+	action,
+	type Model,
+	type ModelConstructor,
+	createModel,
 	Signal,
 	type ReadonlySignal,
 	useSignal,
 	useComputed,
 	useSignalEffect,
+	useModel,
 	untracked,
 };