+++ packages/devtools-adapter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @preact/signals-devtools-adapter
 
+## 0.4.0
+
+### Minor Changes
+
+- [#866](https://github.com/preactjs/signals/pull/866) [`00ba858`](https://github.com/preactjs/signals/commit/00ba858365c135049400607d2f3f380756c96c5d) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Add special case in devtools-ui for component updates
+
+## 0.3.0
+
+### Minor Changes
+
+- [#837](https://github.com/preactjs/signals/pull/837) [`0e0625d`](https://github.com/preactjs/signals/commit/0e0625d12f9a7a7770d2e24c63d5cff7cb422c25) Thanks [@JoviDeCroock](https://github.com/JoviDeCroock)! - Ensure all component/effect dependencies are always visualized in the Graph
+
 ## 0.2.0
 
 ### Minor Changes